ajustes de validação e organização do código

+++ user_register/dynamo/dynamoDB.py
@@ -1,5 +1,6 @@
 import json
 import boto3
+import err.custom_err as err
 
 
 # Cria uma conexão com o serviço DynamoDB
@@ -26,8 +27,13 @@ def get_data():
     return items
     
     
-def post_item(payload):    
-            
+def post_item(payload):
+    
+    response = table.get_item(Key={'id': payload['id']})
+    
+    if 'Item' in response:
+        raise err.ConflictErr(f"ID {payload['id']} is already registered!")
+        
     # Define os valores do item que será inserido
     item = {
         'id': payload['id'],
@@ -42,6 +48,14 @@ def post_item(payload):
     return "User added!"
     
 def update_item(id_user, data):
+    
+    # Verificação se id existe no banco
+    response = table.get_item(Key={'id': id_user})
+    
+    if 'Item' not in response:
+        # O item existe na tabela
+        raise err.NotFoundErr(f"ID {id_user} is not registered!")
+    
     
     # Informe os novos valores para atualizar
     expression_attribute_values = {':age': data['age'], ':email': data['email'], ':userName': data['userName']}
@@ -59,6 +73,13 @@ def update_item(id_user, data):
     return 'User updated'
     
 def delete_user(id_user):
+    
+    # Verificação se id existe no banco
+    response = table.get_item(Key={'id': id_user})
+    
+    if 'Item' not in response:
+        # O item existe na tabela
+        raise err.NotFoundErr(f"ID {id_user} is not registered!")
     
     table.delete_item(
         Key={
